rpc: Move verifier rpc traits to verifier crate.
