+++ core/src/traits/rpc.rs
@@ -6,53 +6,6 @@ use bitcoin::{Address, OutPoint, TxOut, Txid};
 use jsonrpsee::proc_macros::rpc;
 use secp256k1::schnorr;
 
-#[rpc(client, server, namespace = "verifier")]
-pub trait VerifierRpc {
-    #[method(name = "new_deposit")]
-    /// - Check deposit UTXO,
-    /// - Generate random pubNonces, secNonces
-    /// - Save pubNonces and secNonces to a in-memory db
-    /// - Return pubNonces
-    async fn new_deposit(
-        &self,
-        deposit_outpoint: OutPoint,
-        recovery_taproot_address: Address<NetworkUnchecked>,
-        evm_address: EVMAddress,
-    ) -> Result<Vec<MuSigPubNonce>, BridgeError>;
-
-    #[method(name = "operator_kickoffs_generated")]
-    /// - Check the kickoff_utxos
-    /// - for every kickoff_utxo, calculate kickoff2_tx
-    /// - for every kickoff2_tx, partial sign burn_tx (ommitted for now)
-    async fn operator_kickoffs_generated(
-        &self,
-        deposit_outpoint: OutPoint,
-        kickoff_utxos: Vec<UTXO>,
-        operators_kickoff_sigs: Vec<schnorr::Signature>,
-        agg_nonces: Vec<MuSigAggNonce>,
-    ) -> Result<(Vec<MuSigPartialSignature>, Vec<MuSigPartialSignature>), BridgeError>;
-
-    #[method(name = "burn_txs_signed")]
-    /// verify burn txs are signed by verifiers
-    /// sign operator_takes_txs
-    async fn burn_txs_signed(
-        &self,
-        deposit_outpoint: OutPoint,
-        burn_sigs: Vec<schnorr::Signature>,
-        slash_or_take_sigs: Vec<schnorr::Signature>,
-    ) -> Result<Vec<MuSigPartialSignature>, BridgeError>;
-
-    // operator_take_txs_signed
-    #[method(name = "operator_take_txs_signed")]
-    /// verify the operator_take_sigs
-    /// sign move_tx
-    async fn operator_take_txs_signed(
-        &self,
-        deposit_outpoint: OutPoint,
-        operator_take_sigs: Vec<schnorr::Signature>,
-    ) -> Result<MuSigPartialSignature, BridgeError>;
-}
-
 #[rpc(client, server, namespace = "operator")]
 pub trait OperatorRpc {
     #[method(name = "new_deposit")]
